Rename connectWithContext to connectToWebChat
