+++ packages/component/src/Activity/Timestamp.js
@@ -1,9 +1,9 @@
 import React from 'react';
 
-import connectWithContext from '../connectWithContext';
+import connectToWebChat from '../connectToWebChat';
 import TimeAgo from '../Utils/TimeAgo';
 
-export default connectWithContext(
+export default connectToWebChat(
   ({ styleSet }) => ({ styleSet })
 )(
   ({
