+++ packages/component/src/Attachment/AudioAttachment.js
@@ -1,9 +1,9 @@
 import React from 'react';
 
 import AudioContent from './AudioContent';
-import connectWithContext from '../connectWithContext';
+import connectToWebChat from '../connectToWebChat';
 
-export default connectWithContext(
+export default connectToWebChat(
   ({ styleSet }) => ({ styleSet })
 )(
   ({ attachment, styleSet }) =>
