+++ packages/component/src/Attachment/ImageContent.js
@@ -1,9 +1,9 @@
 import React from 'react';
 
-import connectWithContext from '../connectWithContext';
+import connectToWebChat from '../connectToWebChat';
 import CroppedImage from '../Utils/CroppedImage';
 
-export default connectWithContext(
+export default connectToWebChat(
   ({ styleSet }) => ({ styleSet })
 )(
   ({
