+++ packages/component/src/Attachment/TextContent.js
@@ -1,9 +1,9 @@
 import classNames from 'classnames';
 import React from 'react';
 
-import connectWithContext from '../connectWithContext';
+import connectToWebChat from '../connectToWebChat';
 
-export default connectWithContext(
+export default connectToWebChat(
   ({ renderMarkdown, styleSet }) => ({ renderMarkdown, styleSet })
 )(
   ({
