+++ packages/component/src/Attachment/VideoCardAttachment.js
@@ -1,10 +1,10 @@
 import React from 'react';
 
 import CommonCard from './CommonCard';
-import connectWithContext from '../connectWithContext';
+import connectToWebChat from '../connectToWebChat';
 import VideoContent from './VideoContent';
 
-export default connectWithContext(
+export default connectToWebChat(
   ({ styleSet }) => ({ styleSet })
 )(
   ({
