+++ packages/component/src/ErrorBox.js
@@ -1,8 +1,8 @@
 import React from 'react';
 
-import connectWithContext from './connectWithContext';
+import connectToWebChat from './connectToWebChat';
 
-export default connectWithContext(
+export default connectToWebChat(
   ({ styleSet }) => ({ styleSet })
 )(
   ({
