+++ packages/component/src/connectToWebChat.js
